requirements.txt
langchain-core
langchain-community
langgraph
langchain_openai
openai
tzdata
tzlocal
python-dotenv
#pip install "langgraph-cli[inmem]" #this is for local dev server
#pip install debugpy
